2023/04/12

+++ gin_vue_blog_server/model/Model.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type Model struct {
-	ID       uint64    `json:"ID" gorm:"primary_key;auto_increment"`
+	ID       uint      `json:"ID" gorm:"primary_key;auto_increment"`
 	CreateAt time.Time `json:"CreateAt"`
 	UpdateAt time.Time `json:"UpdateAt"`
 }
